+++ src/common/hooks/mutations/useFollow/index.ts
@@ -2,10 +2,10 @@ import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 import { instance } from '~/api';
 
-export const useMutationUnfollow = () => {
-  const followDelete = async ({ id }: { id: string }) => {
+export const useFollow = () => {
+  const followCreate = async ({ id }: { id: string }) => {
     try {
-      await instance.post(`/follow/Delete/${id}`);
+      await instance.post(`/follow/create/${id}`);
     } catch (error) {
       if (error instanceof Error) alert(error.message);
     }
@@ -14,7 +14,7 @@ export const useMutationUnfollow = () => {
   const queryClient = useQueryClient();
 
   const mutation = useMutation({
-    mutationFn: followDelete,
+    mutationFn: followCreate,
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: ['user'] });
     },
